requirements.txt
cloudinary==1.42.2
cryptography==44.0.2
face_recognition==1.3.0
fastapi==0.115.11
ffmpeg_python==0.2.0
imageio==2.37.0
numpy==2.2.4
opencv_python==4.11.0.86
passlib==1.7.4
piexif==1.1.3
Pillow==11.1.0
pydantic==2.10.6
pydantic_settings==2.8.1
python_jose==3.3.0
PyWavelets
SQLAlchemy==2.0.37
sqlmodel==0.0.24
starlette==0.46.1
tqdm==4.67.1
